[unfinished] Moving functionality from Waiter to Kitchen UI and effects for social Fixed: Not creating new posts
