+++ Waiter/views/admin/connect.twig
@@ -85,7 +85,13 @@
                     <label for="kitchen_api">
                         Kitchen Api
                     </label>
-                    <input type="text" class="persist" name="SoupWaiter.kitchen_user" id="kitchen_api" value="{{ soup.kitchen_api }}"/>
+                    <input type="text" class="persist" name="SoupWaiter.kitchen_api" id="kitchen_api" value="{{ soup.kitchen_api }}"/>
+                </div>
+                <div class="field">
+                    <label for="social_api">
+                        Social Api
+                    </label>
+                    <input type="text" class="persist" name="SoupWaiter.social_api" id="social_api" value="{{ soup.social_api }}"/>
                 </div>
             </div>
         </div>
