+++ Waiter/views/admin/dash.twig
@@ -1,10 +1,17 @@
-{% include 'admin/admin-tabs.twig' %}
-<h1>Soup Vacation Dashboard</h1>
-<div class="row">
-    <div class="col-xs-12 col-sm-6 col-lg-4">
-        I'm in a column
-    </div>
-    <div class="col-xs-12 col-sm-6 col-lg-4">
-        And so am I
+{% extends 'admin/vacation-soup.twig' %}
+{% block content %}
+{% embed 'admin/form.twig' %}
+{% block form_name %}dash{% endblock form_name %}
+{% block form_content %}
+<div class="vs-content tab-dash row">
+    <div class="col-lg-8 col-sm-12">
+        <div class="vs-panel row">
+            <div id="title" class="col-sm-12">
+                <h1>Dashboard</h1>
+            </div>
+        </div>
     </div>
 </div>
+{% endblock form_content %}
+{% endembed %}
+{% endblock content %}
