+++ Waiter/views/admin/form.twig
@@ -1,30 +1,52 @@
 <script>
     jQuery(function() {
         jQuery('.persist').change(function () {
+            var ajaxCall;
             var jElem = jQuery(this);
             this.disabled=true;
 
             var value = jElem.val();
             if ('checkbox'===this.type){
-                value = (this.checked)?1:0; // otherwise boolean strings of 'false'
+                value = (this.checked)?1:0; // otherwise boolean strings of 'false' handled as true in php
             }
 
-            var data = {
-                action: "soup",
-                tab: "{{ tab }}",
-                _vs_nonce: jQuery('INPUT[name="_vs_nonce"]').val(),
-                name: jElem.attr("name"),
-                value: value,
-                id: jElem.prop("id")
-            };
+            var broken = jElem.attr("name").split('.');
+            var type = broken[0];
+            var name = broken[1];
 
-            jQuery.ajax({
-                type: "post",
-                dataType: "json",
-                context: this,
-                url: ajaxurl,
-                data: data
-            }).fail(function(jqXHR,status,error) {
+            // If SoupWaiter or SoupWaiterAdmin send as current
+            // Else send to Kitchen, using new API
+
+            if (true || "SoupWaiter"==type || "SoupWaiterAdmin"==type){
+
+                ajaxCall = {
+                    type: "post",
+                    dataType: "json",
+                    context: this,
+                    url: ajaxurl,
+                    data: {
+                        action: "soup",
+                        tab: "{{ tab }}",
+                        _vs_nonce: jQuery('INPUT[name="_vs_nonce"]').val(),
+                        name: jElem.attr("name"),
+                        value: value,
+                        id: jElem.prop("id")
+                    }
+                };
+            } else {
+                ajaxCall = {
+                    type: "post",
+                    dataType: "json",
+                    context: this,
+                    url: '{{ soup.social_api~'/' }}'+type,
+                    data: {
+                        name: name,
+                        value: value,
+                    }
+                };
+            }
+
+            jQuery.ajax(ajaxCall).fail(function(jqXHR,status,error) {
                 console.log('fail (vacation-soup) '+status+' '+error);
                 alert('Error: Could not save change, refreshing page may help\n'+status);
             }).done(function(response) {
