+++ Waiter/views/social/NotImplemented.twig
@@ -1,6 +1,6 @@
 <div class="field">
     <p>
-        We have not yet implemented auto-posting to {{ soc.display_name }}. Voting is set to {{ soc.vote }}.
+        We have not yet implemented auto-posting to {{ soc.display_name }}.
         If you would like to connect to {{ soc.display_name }}, please vote for it by ticking the checkbox, and we will move it up the list.
     </p>
     <label for="vote_{{ key|lower }}">Vote for {{ soc.display_name }}</label>
